pyproject.toml
[project]
name = "LLM2Deck"
version = "0.1.2"
description = "Convert LLM output to Anki decks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.3.0",
    "browser-cookie3>=0.20.1",
    "cerebras-cloud-sdk>=1.59.0",
    "g4f[all]>=6.7.2",
    "gemini-webapi>=1.17.3",
    "genanki>=0.13.1",
    "google-genai>=1.33.0",
    "markdown>=3.10",
    "openai>=2.11.0",
    "pillow>=11.3.0",
    "pydantic>=2.12.5",
    "pygments>=2.19.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=13.0.0",
    "selenium>=4.35.0",
    "sqlalchemy>=2.0.0",
    "zendriver>=0.14.0",
]

[project.scripts]
llm2deck = "src.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require API keys)",
]
# Exclude experiments directory (contains runnable scripts, not tests)
norecursedirs = ["tests/experiments"]
pythonpath = ["."]
